[Feat] 풍향,풍속 시스템 테스트 완료
풍향,풍속 시스템 연동 및 테스트 완료

+++ Assets/PMS/PMS_Scripts/WindData.cs
@@ -6,9 +6,9 @@ using UnityEngine;
 public struct WindData
 {
     public Vector3 direction;
-    public float speed;
+    public int speed;
 
-    public WindData(Vector3 dir, float spd)
+    public WindData(Vector3 dir, int spd)
     {
         direction = dir;
         speed = spd;
